chore(data files): create endpoint class to download method & add tqdm

+++ pysus/ftp/__init__.py
@@ -8,6 +8,8 @@ from typing import Any, Dict, List, Optional, Self, Set, Union
 import humanize
 from aioftp import Client
 from loguru import logger
+from pysus.data.local import Data
+from tqdm import tqdm
 
 CACHEPATH = os.getenv(
     "PYSUS_CACHEPATH", os.path.join(str(pathlib.Path.home()), "pysus")
@@ -84,21 +86,50 @@ class File:
         info["modify"] = self.__info__["modify"].strftime("%Y-%m-%d %I:%M%p")
         return info
 
-    def download(self, local_dir: str = CACHEPATH) -> str:
+    def download(self, local_dir: str = CACHEPATH, _pbar=None) -> Data:
         _dir = pathlib.Path(local_dir)
         _dir.mkdir(exist_ok=True, parents=True)
         filepath = _dir / self.basename
+        filesize = int(self.__info__["size"])
+
+        if _pbar:
+            _pbar.unit = "B"
+            _pbar.unit_scale = True
+            _pbar.reset(total=filesize)
+
+        _parquet = filepath.with_suffix(".parquet")
+        if _parquet.exists():
+            if _pbar:
+                _pbar.update(filesize - _pbar.n)
+            return Data(str(_parquet), _pbar=_pbar)
+
+        _dbf = filepath.with_suffix(".dbf")
+        if _dbf.exists():
+            if _pbar:
+                _pbar.update(filesize - _pbar.n)
+            return Data(str(_dbf), _pbar=_pbar)
 
         if filepath.exists():
-            return str(filepath)
+            if _pbar:
+                _pbar.update(filesize - _pbar.n)
+            return Data(str(filepath), _pbar=_pbar)
+
+        if _pbar:
+            _pbar.set_description(f"{self.basename}")
 
         try:
             ftp = ftp = FTP("ftp.datasus.gov.br")
             ftp.login()
             output = open(f"{filepath}", "wb")
+
+            def callback(data):
+                output.write(data)
+                if _pbar:
+                    _pbar.update(len(data))
+
             ftp.retrbinary(
                 f"RETR {self.path}",
-                output.write,
+                callback,
             )
         except Exception as exc:
             raise exc
@@ -106,10 +137,11 @@ class File:
             ftp.close()
             output.close()
 
-        logger.info(f"{self.basename} downloaded at {local_dir}")
-        return str(filepath)
+        if _pbar:
+            _pbar.update(filesize - _pbar.n)
+        return Data(str(filepath), _pbar=_pbar)
 
-    async def async_download(self, local_dir: str = CACHEPATH) -> None:
+    async def async_download(self, local_dir: str = CACHEPATH) -> Data:
         # aioftp.Client.parse_list_line_custom
         def line_file_parser(file_line):
             line = file_line.decode("utf-8")
@@ -141,16 +173,25 @@ class File:
             else local_dir + "/" + str(self.basename)
         )
 
+        _parquet = filepath.with_suffix(".parquet")
+        if _parquet.exists():
+            return Data(str(_parquet))
+
+        _dbf = filepath.with_suffix(".dbf")
+        if _dbf.exists():
+            return Data(str(_dbf))
+
         if filepath.exists():
-            logger.debug(output)
-        else:
-            async with Client.context(
-                host="ftp.datasus.gov.br",
-                parse_list_line_custom=line_file_parser,
-            ) as client:
-                await client.login()
-                await client.download(self.path, output, write_into=True)
-                logger.debug(output)
+            return Data(output)
+
+        async with Client.context(
+            host="ftp.datasus.gov.br",
+            parse_list_line_custom=line_file_parser,
+        ) as client:
+            await client.login()
+            await client.download(self.path, output, write_into=True)
+
+        return Data(output)
 
 
 CACHE: Dict = {}
@@ -471,10 +512,15 @@ class Database:
         """
         Downloads a list of Files.
         """
+        files = to_list(files)
+        pbar = tqdm(total=len(files), dynamic_ncols=True)
         dfiles = []
         for file in files:
             if isinstance(file, File):
-                dfiles.append(file.download(local_dir=local_dir))
+                dfiles.append(file.download(local_dir=local_dir, _pbar=pbar))
+        pbar.close()
+        if len(dfiles) == 1:
+            return dfiles[0]
         return dfiles
 
     async def async_download(
